modificado arquivo _MODELO em htm para profile.php, mudado nome dos arquivos: 'LoginHndler' para 'UsuarioHandler'e 'public/media/covers/cover.jpg' para 'public/media/covers/fotoCapa.jpg'
