+++ _MODELO/perfil.html
@@ -108,7 +108,7 @@
             <div class="row">
                 <div class="box flex-1 border-top-flat">
                     <div class="box-body">
-                        <div class="profile-cover" style="background-image: url('media/covers/cover.jpg');"></div>
+                        <div class="profile-cover" style="background-image: url('media/covers/fotoCapa.jpg');"></div>
                         <div class="profile-info m-20 row">
                             <div class="profile-info-avatar">
                                 <img src="media/avatars/fotoPerfil.jpg" />
